fixed the database mismatch, now its working correctly

+++ frontend/app/src/main/java/com/example/frontendinternship/domain/model/ProductWithCount.kt
@@ -9,6 +9,6 @@ data class ProductWithCount(
 ) {
 }
 
-fun ProductWithCount.getCost(): Float {
+fun ProductWithCount.getCost(): Double {
     return product.price * count
 }
