+++ frontend/app/src/main/java/com/example/frontendinternship/domain/model/VatTotalAmounts.kt
@@ -1,8 +1,8 @@
 package com.example.frontendinternship.domain.model
 
-data class VatTotalAmounts (
-    var amountVat0: Float = 0f,
-    var amountVat1: Float = 0f,
-    var amountVat10: Float = 0f,
-    var amountVat20: Float = 0f,
+data class VatTotalAmounts(
+    var amountVat0: Double = 0.0,
+    var amountVat1: Double = 0.0,
+    var amountVat10: Double = 0.0,
+    var amountVat20: Double = 0.0,
 )
